Extracted MultiValuedBinaryDocValuesField into a separate class and removed INSERTION order
